show thumbnails on preview page to allow easy browsing
Only shows the first 16 thumbnails (a number you can easily change).
Generating the thumbnails for the first time takes a bit of time,
but they are cached for a month after being generated.

We'll show thumbnails only for the most recent images.

You need to install php5-gd for this to work:
apt-get install php5-gd