RPi_Cam_Cmd: handle remote calls to the cam
I wanted to use a simple 'API' that I could use to have other devices send commands to the cam. As it seems, sending multiple commands at the same time (or while a previous one is still processing) seems to crash the Pi.

To solve this, I created this wrapper script that 'locks' the cam and does not allow new commands to execute when a command is already executing. In my case, all commands are sent by my Fibaro Home Center zwave controller using this wrapper script.